Search function now works. "app" is changed to "main".

+++ Project/main/forms.py
@@ -5,4 +5,4 @@ class CreateNewAccount(forms.Form):
     password = forms.CharField(max_length=200)
     email = forms.EmailField()
     role = forms.IntegerField() #0 = no account 1 = buyer 2 = seller 3 = admin
-    balance = forms.IntegerField()
+    balance = forms.FloatField()
